// front/src/app/swagger/model/categoryResponseDto.ts
/**
 * ExamenTec API
 *
 * Contact: support@examentec.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CategoryResponseDto { 
    id?: string;
    name?: string | null;
    description?: string | null;
    createdDate?: string;
}